separate body and head plugins entirely
